pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aiforge-memory"
version = "0.1.0"
description = "AiForgeMemory — unified code memory: Repo / Service / File / Symbol + chunk vectors. Repomix + tree-sitter + LLM + Neo4j."
requires-python = ">=3.11"
authors = [{name = "Manikanta Reddy Pasala"}]
license = {text = "MIT"}
readme = "README.md"
dependencies = [
  "neo4j>=5.20",
  "openai>=1.50",
  "httpx>=0.27",
  "PyYAML>=6.0",
  "tree-sitter>=0.25",
  "tree-sitter-language-pack>=0.13",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.4",
]

[project.scripts]
aiforge-memory = "aiforge_memory.api.cli:main"

[tool.setuptools.packages.find]
include = ["aiforge_memory*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["aiforge_memory/tests"]
addopts = "-ra"
markers = [
    "live_neo4j: requires Neo4j running at AIFORGE_NEO4J_URI",
    "live_llm: requires planner LLM at AIFORGE_INTENT_LM_URL",
    "live_repomix: requires `repomix` binary on PATH",
]
